заменил весь BigDecimal в проекте на Double

+++ src/main/java/com/overridetech/funpay_monitor/dto/CurrencyRateDto.java
@@ -5,12 +5,11 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @NoArgsConstructor
 @Getter
 @Setter
 public class CurrencyRateDto {
-    private BigDecimal usdrub;
-    private BigDecimal rubusd;
+    private Double usdrub;
+    private Double rubusd;
 }
